HOS-24417 Update mlo related parameters
