+++ plugins/inputs/ah_trap/ah_trap.go
@@ -118,45 +118,7 @@ func (t *TrapPlugin) Gather_Ah_Logen(trap AhTrapMsg, acc telegraf.Accumulator) e
                 rawSize := int(unsafe.Sizeof(ahconnectionchange))
                 copy((*[1 << 10]byte)(unsafe.Pointer(&ahconnectionchange))[:rawSize], trap.Union[:rawSize])
 
-                acc.AddFields("TrapEvent", map[string]interface{}{
-                        "trapObjName_connectionChangeTrap":          ahutil.CleanCString(ahconnectionchange.Name[:]),
-			"ssid_connectionChangeTrap":              ahutil.CleanCString(ahconnectionchange.Ssid[:]),
-			"hostName_connectionChangeTrap":          ahutil.CleanCString(ahconnectionchange.HostName[:]),
-			"userName_connectionChangeTrap":          ahutil.CleanCString(ahconnectionchange.UserName[:]),
-			"objectType_connectionChangeTrap":        ahconnectionchange.ObjectType,
-			"remoteId_connectionChangeTrap":          ahutil.FormatMac(ahconnectionchange.RemoteID),
-			"bssid_connectionChangeTrap":             ahutil.FormatMac(ahconnectionchange.BSSID),
-			"curState_connectionChangeTrap":          ahconnectionchange.CurState,
-			"clientIp_connectionChangeTrap":          ahutil.IntToIpv4(ahconnectionchange.ClientIP),
-			"clientAuthMethod_connectionChangeTrap":  ahconnectionchange.ClientAuthMethod,
-			"clientEncryptMethod_connectionChangeTrap": ahconnectionchange.ClientEncryptMethod,
-			"clientMacProto_connectionChangeTrap":    ahconnectionchange.ClientMacProto,
-			"clientVlan_connectionChangeTrap":        ahconnectionchange.ClientVLAN,
-			"clientUpId_connectionChangeTrap":        ahconnectionchange.ClientUPID,
-			"clientChannel_connectionChangeTrap":     ahconnectionchange.ClientChannel,
-			"clientCwpUsed_connectionChangeTrap":     ahconnectionchange.ClientCWPUsed,
-			"associationTime_connectionChangeTrap":   ahconnectionchange.AssociationTime,
-			"ifIndex_keys_connectionChangeTrap":      ahconnectionchange.IfIndex,
-                        "name_keys_connectionChangeTrap":         ahutil.CleanCString(ahconnectionchange.IfName[:]),
-			"rssi_connectionChangeTrap":              ahconnectionchange.RSSI,
-			"snr_connectionChangeTrap":               ahconnectionchange.SNR,
-			"profile_connectionChangeTrap":           ahutil.CleanCString(ahconnectionchange.ProfName[:]),
-			"authUsed_connectionChangeTrap":          ahconnectionchange.ClientMacBasedAuthUsed,
-			"os_connectionChangeTrap":                ahutil.CleanCString(ahconnectionchange.OS[:]),
-			"option55_connectionChangeTrap":          ahutil.CleanCString(ahconnectionchange.Option55[:]),
-			"mgtStatus_connectionChangeTrap":         ahconnectionchange.MgtStus,
-			"staAddr6Num_connectionChangeTrap":       ahconnectionchange.StaAddr6Num,
-			"staAddr6_connectionChangeTrap":          intToIPv6(ahconnectionchange.StaAddr6[:], int(ahconnectionchange.StaAddr6Num)),
-			"deauthReason_connectionChangeTrap":      ahconnectionchange.DeauthReason,
-			"roamTime_connectionChangeTrap":          ahconnectionchange.RoamTime,
-			"assocTime_connectionChangeTrap":         ahconnectionchange.AssocTime,
-			"authTime_connectionChangeTrap":          ahconnectionchange.AuthTime,
-			"radioProfile_connectionChangeTrap":      ahutil.CleanCString(ahconnectionchange.RadioProf[:]),
-			"negotiateKbps_connectionChangeTrap":     ahconnectionchange.NegotiateKbps,
-                        "severityLevel_trapMessage_connectionChangeTrap": trap.Level,
-                        "msgId_trapMessage_connectionChangeTrap": trap.MsgID,
-                        "desc_trapMessage_connectionChangeTrap":  ahutil.CleanCString(trap.Desc[:]),
-                }, nil)
+		gatherConnectionChangeTrap(ahconnectionchange, trap, acc)
 
 	case AH_IDP_AP_EVENT_TRAP_TYPE:
                 var ahidpapevent AhIdpApEventTrap
